requirements.txt
graphviz==0.7.1
h5py==2.8.0
h5py-cache==1.0
numpy==1.14.6
Pillow==4.2.1
requests==2.18.4
progressbar2==3.34.3
tabulate==0.8.1
scipy==1.1.0
torch==0.3.1.post3
torchvision==0.2.0
matplotlib==2.1.0
seaborn==0.9.0
pytest==3.8.2
git+https://github.com/pytorch/tnt.git@511db2bcdf87f12aacca46726778b8ee73cef0e4#egg=torchnet
git+https://github.com/anibali/pyshowoff.git@cfef2201547cbe3f8fb1e74ff62336d8d397d1dd#egg=pyshowoff
git+https://github.com/anibali/tele.git@244b103ed826bb135eea857e7541099ffb47e135#egg=tele
